Adding options for overwriting

+++ mtslib.py
@@ -5,6 +5,7 @@ import os
 import os.path
 import glob
 import sys
+import shutil
 import cmn.cmn as cmn
 
 
@@ -16,6 +17,7 @@ def sortmtsfile(mtsfile, wingdetdir):
     root, ext = os.path.splitext(os.path.basename(mtsfile))
     newdir = cmn.makenewdir(os.path.join(wingdetdir, root))
     os.rename(mtsfile, os.path.join(newdir, root+ext))
+
 
 def sortmtsdir(params):
     """Run in a directory with multiple MTS files. Moves each MTS file to a new 
@@ -59,31 +61,56 @@ def convimg(infile, ext='tif'):
         sys.exit(0)
    
 
-def convmovie(mtsfile, outfile, start, dur, specdur, ext='tif'):
+def convmovie(mtsfile, outfile, start, dur, specdur, ext='tif', overwrite='no'):
     """Converts a single MTS file to a series of images using ffmpeg.
     Inputs:
     mtsfile - name of MTS file
     outfile - name of output file, including extension
     start - time to start conversion, in seconds
     dur - duration of movie to be converted, in seconds
-    ext - type of image file (default is tif).
+    ext - type of image file (default is tif)
+    overwrite - 'yes' or 'no'; if no, then the script will exit if the movie 
+    folder already exists
     """
     
+    if os.path.exists('movie') and overwrite == 'no':
+        sys.exit('Already converted.')
+    if os.path.exists('movie') and overwrite == 'yes':
+        shutil.rmtree('movie')
+    print('Converting MTS file to avi')
     mtsconv(mtsfile, outfile, start, dur, specdur)
     cmn.makenewdir('movie')
     os.rename(outfile, 'movie/'+outfile)
     os.chdir('movie')
+    print('Converting avi to image sequence')
     convimg(outfile, ext)
     os.remove(outfile)
 
 
-def convmovies(fdir, start, dur, specdur):
+def convmovies(fdir, start, dur, specdur, boverwrite):
+    """Converts MTS files in the moviexxx directories to a series of images 
+    using ffmpeg.
+    Inputs:
+    fdir - expt/ directory
+    start - time to start conversion, in seconds
+    dur - duration of movie to be converted, in seconds
+    ext - type of image file (default is tif)
+    boverwrite - 'yes' or 'no'; if 'no', then the for loop will continue to the 
+    next iteration if the movie/ folder already exists
+    """
+    
     dirs = cmn.listsortfs(fdir)
     for d in dirs:
         os.chdir(d)
+        print(d)
         mtsfile = glob.glob('*{0}'.format('MTS'))[0]
         outfile = os.path.splitext(mtsfile)[0] + '.avi'
-        print(os.getcwd())
+        movdir = os.path.join(d, 'movie')
+        if os.path.exists(movdir) and boverwrite == 'no':
+            print('Already converted.')
+            continue
+        if os.path.exists(movdir) and boverwrite == 'yes':
+            shutil.rmtree(movdir)
         convmovie(mtsfile, outfile, start, dur, specdur)
